fix: moved test method into a /test directory
