+++ src/main/java/com/seailz/discordjv/DiscordJv.java
@@ -167,28 +167,6 @@ public class DiscordJv {
         this(token, intents, APIVersion.getLatest());
     }
 
-    // This method is used for testing purposes only
-    public static void main(String[] args) throws ExecutionException, InterruptedException, IOException {
-        String token = "";
-        File file = new File("token.txt");
-        // get first line of that file
-        try (FileReader reader = new FileReader(file)) {
-            token = new BufferedReader(reader).readLine();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        DiscordJv discordJv = new DiscordJv(token, EnumSet.of(Intent.GUILDS, Intent.GUILD_MESSAGES));
-
-        ArrayList<Activity> activities = new ArrayList<>();
-        activities.add(
-                new Activity("Hello World2", ActivityType.WATCHING)
-        );
-        Status status = new Status(0, activities.toArray(new Activity[0]), StatusType.DO_NOT_DISTURB, false);
-        discordJv.setStatus(status);
-
-        discordJv.clearCommands();
-    }
-
     /**
      * Stops the bot from shutting down when processes are finished
      */
